Added Finished Laser Beam
Added laser beam particles that appear anytime the laser collides with an object. This prompts the Laser Beam particle effect to activate three sub emitters. One that creates a dissipation of energy on the object being struck by the laser, one for the steam effect rising off of the surface being hit, and another for the sparks flying out of the dissipation of energy.